# packages/contracts/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "AuthStatus": {
        "additionalProperties": false,
        "properties": {
          "authenticated": {
            "title": "Authenticated",
            "type": "boolean"
          },
          "mode": {
            "enum": [
              "disabled",
              "owner"
            ],
            "title": "Mode",
            "type": "string"
          },
          "owner": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OwnerIdentity"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "mode",
          "authenticated"
        ],
        "title": "AuthStatus",
        "type": "object"
      },
...
      "DataMode": {
        "description": "Origine isolée des données applicatives.",
        "enum": [
          "mock",
          "real"
        ],
        "title": "DataMode",
        "type": "string"
      },
...
      "DependencyStatus": {
        "additionalProperties": false,
        "description": "État synthétique d'une dépendance de disponibilité.",
        "properties": {
          "reasonCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoncode"
          },
          "status": {
            "enum": [
              "available",
              "unavailable"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "DependencyStatus",
        "type": "object"
      },
...
      "GateStatus": {
        "enum": [
          "NO-GO",
          "GO"
        ],
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OperationalStatus": {
        "additionalProperties": false,
        "properties": {
          "backups": {
            "$ref": "#/components/schemas/BackupOperationalHealth"
          },
          "jobCounts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Jobcounts",
            "type": "object"
          },
          "mappingBacklog": {
            "minimum": 0.0,
            "title": "Mappingbacklog",
            "type": "integer"
          },
          "metrics": {
            "$ref": "#/components/schemas/OperationalMetrics"
          },
          "model": {
            "$ref": "#/components/schemas/ModelOperationalHealth"
          },
          "readsAvailable": {
            "title": "Readsavailable",
            "type": "boolean"
          },
          "source": {
            "$ref": "#/components/schemas/SourceOperationalHealth"
          }
        },
        "required": [
          "readsAvailable",
          "source",
          "model",
          "mappingBacklog",
          "jobCounts",
          "backups",
          "metrics"
        ],
        "title": "OperationalStatus",
        "type": "object"
      },
...
      "OwnerIdentity": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username"
        ],
        "title": "OwnerIdentity",
        "type": "object"
      },
      "OwnerLoginRequest": {
        "additionalProperties": false,
        "properties": {
          "password": {
            "format": "password",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Password",
            "type": "string",
            "writeOnly": true
          },
          "username": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "OwnerLoginRequest",
        "type": "object"
      },
...
      "ProblemDetails": {
        "additionalProperties": false,
        "description": "Erreur HTTP compatible RFC 9457, successeur de RFC 7807.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "context": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Context",
            "type": "object"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "detail",
          "instance",
          "code"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
      "ReleaseAudience": {
        "enum": [
          "personal",
          "public",
          "commercial"
        ],
        "type": "string"
      },
      "ReleaseCompliance": {
        "additionalProperties": false,
        "properties": {
          "audience": {
            "$ref": "#/components/schemas/ReleaseAudience"
          },
          "gates": {
            "additionalProperties": {
              "$ref": "#/components/schemas/GateStatus"
            },
            "title": "Gates",
            "type": "object"
          },
          "publicReleaseAllowed": {
            "title": "Publicreleaseallowed",
            "type": "boolean"
          },
          "stakeProviderEnabled": {
            "const": false,
            "default": false,
            "title": "Stakeproviderenabled",
            "type": "boolean"
          }
        },
        "required": [
          "audience",
          "gates",
          "publicReleaseAllowed"
        ],
        "title": "ReleaseCompliance",
        "type": "object"
      },
...
      "SystemStatusResponse": {
        "additionalProperties": false,
        "description": "État minimal versionné du système.",
        "properties": {
          "apiVersion": {
            "title": "Apiversion",
            "type": "string"
          },
          "dataMode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "dependencies": {
            "additionalProperties": {
              "$ref": "#/components/schemas/DependencyStatus"
            },
            "title": "Dependencies",
            "type": "object"
          },
          "generatedAt": {
            "format": "date-time",
            "title": "Generatedat",
            "type": "string"
          },
          "operations": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OperationalStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "enum": [
              "ready",
              "degraded"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "apiVersion",
          "dataMode",
          "generatedAt",
          "dependencies"
        ],
        "title": "SystemStatusResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Metiquo API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnerLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/system/compliance": {
      "get": {
        "operationId": "release_compliance_api_v1_system_compliance_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseCompliance"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Ressource introuvable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Requête invalide"
          }
        },
        "summary": "Release Compliance",
        "tags": [
          "system"
        ]
      }
    },
    "/api/v1/system/status": {
      "get": {
        "operationId": "system_status_api_v1_system_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Ressource introuvable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {}
            },
            "description": "Requête invalide"
          }
        },
        "summary": "System Status",
        "tags": [
          "system"
        ]
      }
    },
...
  }
}